# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mukmin Life API",
        "contact": {
            "name": "Mukmin Life API Support",
            "email": "hello@mukmin.life"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/v2",
    "paths": {
        "/prayer_time/{zone}/{date}": {
            "get": {
                "description": "get prayer time by date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List prayer times of the day",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Zone",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/prayer_time.PrayerTimeOutput"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "prayer_time.PrayerTimeOutput": {
            "type": "object",
            "properties": {
                "asr": {
                    "type": "string"
                },
                "dhuhr": {
                    "type": "string"
                },
                "fajr": {
                    "type": "string"
                },
                "imsak": {
                    "type": "string"
                },
                "isha": {
                    "type": "string"
                },
                "maghrib": {
                    "type": "string"
                },
                "syuruk": {
                    "type": "string"
                }
            }
        }
    }
}